Moved tank class and subclasses from nodes module to their own module
